feat: make eomt installable as a package
