+++ eomt/training/mask_classification_panoptic.py
@@ -1,27 +1,24 @@
 # ---------------------------------------------------------------
 # © 2025 Mobile Perception Systems Lab at TU/e. All rights reserved.
 # Licensed under the MIT License.
-#
-# Portions of this file are adapted from the Mask2Former repository
-# by Facebook, Inc. and its affiliates, used under the Apache 2.0 License.
 # ---------------------------------------------------------------
 
 
 from typing import List, Optional
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from training.mask_classification_loss import MaskClassificationLoss
-from training.lightning_module import LightningModule
+from eomt.training.mask_classification_loss import MaskClassificationLoss
+from eomt.training.lightning_module import LightningModule
 
 
-class MaskClassificationInstance(LightningModule):
+class MaskClassificationPanoptic(LightningModule):
     def __init__(
         self,
         network: nn.Module,
         img_size: tuple[int, int],
         num_classes: int,
+        stuff_classes: list[int],
         attn_mask_annealing_enabled: bool,
         attn_mask_annealing_start_steps: Optional[list[int]] = None,
         attn_mask_annealing_end_steps: Optional[list[int]] = None,
@@ -41,7 +38,6 @@ class MaskClassificationInstance(LightningModule):
         class_coefficient: float = 2.0,
         mask_thresh: float = 0.8,
         overlap_thresh: float = 0.8,
-        eval_top_k_instances: int = 100,
         ckpt_path: Optional[str] = None,
         delta_weights: bool = False,
         load_ckpt_class_head: bool = True,
@@ -69,8 +65,7 @@ class MaskClassificationInstance(LightningModule):
 
         self.mask_thresh = mask_thresh
         self.overlap_thresh = overlap_thresh
-        self.stuff_classes: List[int] = []
-        self.eval_top_k_instances = eval_top_k_instances
+        self.stuff_classes = stuff_classes
 
         self.criterion = MaskClassificationLoss(
             num_points=num_points,
@@ -83,7 +78,12 @@ class MaskClassificationInstance(LightningModule):
             no_object_coefficient=no_object_coefficient,
         )
 
-        self.init_metrics_instance(self.network.num_blocks + 1 if self.network.masked_attn_enabled else 1)
+        thing_classes = [i for i in range(num_classes) if i not in stuff_classes]
+        self.init_metrics_panoptic(
+            thing_classes,
+            stuff_classes,
+            self.network.num_blocks + 1 if self.network.masked_attn_enabled else 1,
+        )
 
     def eval_step(
         self,
@@ -97,6 +97,9 @@ class MaskClassificationInstance(LightningModule):
         transformed_imgs = self.resize_and_pad_imgs_instance_panoptic(imgs)
         mask_logits_per_layer, class_logits_per_layer = self(transformed_imgs)
 
+        is_crowds = [target["is_crowd"] for target in targets]
+        targets = self.to_per_pixel_targets_panoptic(targets)
+
         for i, (mask_logits, class_logits) in enumerate(
             list(zip(mask_logits_per_layer, class_logits_per_layer))
         ):
@@ -104,50 +107,17 @@ class MaskClassificationInstance(LightningModule):
             mask_logits = self.revert_resize_and_pad_logits_instance_panoptic(
                 mask_logits, img_sizes
             )
-
-            preds, targets_ = [], []
-            for j in range(len(mask_logits)):
-                scores = class_logits[j].softmax(dim=-1)[:, :-1]
-                labels = (
-                    torch.arange(scores.shape[-1], device=self.device)
-                    .unsqueeze(0)
-                    .repeat(scores.shape[0], 1)
-                    .flatten(0, 1)
-                )
-
-                topk_scores, topk_indices = scores.flatten(0, 1).topk(
-                    self.eval_top_k_instances, sorted=False
-                )
-                labels = labels[topk_indices]
-
-                topk_indices = topk_indices // scores.shape[-1]
-                mask_logits[j] = mask_logits[j][topk_indices]
-
-                masks = mask_logits[j] > 0
-                mask_scores = (
-                    mask_logits[j].sigmoid().flatten(1) * masks.flatten(1)
-                ).sum(1) / (masks.flatten(1).sum(1) + 1e-6)
-                scores = topk_scores * mask_scores
-
-                preds.append(
-                    dict(
-                        masks=masks,
-                        labels=labels,
-                        scores=scores,
-                    )
-                )
-                targets_.append(
-                    dict(
-                        masks=targets[j]["masks"],
-                        labels=targets[j]["labels"],
-                        iscrowd=targets[j]["is_crowd"],
-                    )
-                )
-
-            self.update_metrics_instance(preds, targets, i)
+            preds = self.to_per_pixel_preds_panoptic(
+                mask_logits,
+                class_logits,
+                self.stuff_classes,
+                self.mask_thresh,
+                self.overlap_thresh,
+            )
+            self.update_metrics_panoptic(preds, targets, is_crowds, i)
 
     def on_validation_epoch_end(self):
-        self._on_eval_epoch_end_instance("val")
+        self._on_eval_epoch_end_panoptic("val")
 
     def on_validation_end(self):
-        self._on_eval_end_instance("val")
+        self._on_eval_end_panoptic("val")
